+++ HISTORY.rst
@@ -4,7 +4,7 @@ History
 
 v0.10.0 (NEXT)
 -------------
-* Add Quantal linear as a default model instead of Mulstage-1
+* Add Quantal linear as a default model instead of Multistage-1
 * Update documentation on default model recommendation logic
 * Fix special case where all BMDLs are 0 for model recommendations